Make std::initializer_list visible.
With C++11 uniform initialization and  initializer lists we cannot
longer easily see whether we are calling a normal constructor or one
which takes a initializer_list. For example:

struct V {
  V(int x) {}
  V(std::initializer_list<int> x) {}

};

V v{1};
V vv{1, 2};

Surprisingly to some people, in both cases the second constructor, taking
a std::initializer_list, is used.

This change shows the difference by putting the name
std::initializer_list<T> in front of an initializer list.