make lambda name uniqu(er) by adding the column number to the name.
As a lambda has no name the system was to create one with __lambda_ as
start and the line number in which the lambda appears. A lambda in line
5 became: __lambda_5. Now, there are multiple ways to get two ore more
lambdas within the same line. For example with the comma operator (see:
tests/MultipleLambdasInOneLineTest.cpp). The system is obviously broken.
To stay unique the column number is added in addition. The lambda in
line 5 does now become: __lambda_5_4, assuming it is declared in column
4.

The idea of the whole system is to keep the name of a lambda stable
between two runs. Other ways would be using the address or the ODR hash.
With the column number the system should be stable and unique..