Added macOS to travis and enable code coverage.
Added macOS as a travis CI target by using pre-build clang and LLVM
binaries. The normal clang installation isn't enough, as the LLVM
libraries are missing. In addition enable the tests for the macOS build.

Further, code coverage is done and uploaded to codecov.io using linux
and g++-7. Sadly lcov is not able to handle the updated format g++-8
creates. To skip additional installs g++-7 is allowed as minimum
version.