Fixed #13: const disappeared in case of auto.
Due to lambdas there is special handling for QualType in case it is a
RecordDecl or a ClassTemplateSpecializationDecl. It appears that, if the
type is a ReferenceType the CV qualifiers come from the reference type
getPointeeTypeAsWritten(). Ohterwise they are empty. This fix adds
special handling for that, nearly the same way as TypePrinter.cpp does
it.