Added support for expansion of class templates.
Up to now class templates ended up with just the class tag and name.
This change expands the whole class, including methods and variables.
There are some cases where a method has no body in the expanded version.
In that case this method is unused in the instatiated version.

This changes also brings support for more statements/expressions. Like
the placement delete or inline variable initialization.

A further simplification is the attempt to introduce a function which is
responsible to adding parens or curlys.