Major rewrite to catch even more statements.
Initially a lot of AST matchers where involved to match certain
hand-picked statements (like range-based for loops, auto or
if-statements). With the evolution C++ Insights this became a burden as
all the matchers needed to be locked against each other. Failing here
lead to a couple of issues. Another drawback is, that there are code
parts which remained untouched.

Now, this patch removes most of these AST matchers and instead matches
FunctionsDecl's. This should rewrite all functions, classes and function
/ class templates. With that C++ Insights should have a much broader
rewrite coverage. However, there is another disadvantage. The code does
look less beautiful as some newlines are missing. clang-format can
correct these on local installations so for now it is accepted.