Fix #50: Remove second and invalid ampersand for an l-value ref.
It looks like clang does handle l-value references in the AST dump
wrong. It adds a second ampersand which makes it an r-value reference.
As #50 points out this second ampersand leads to invalid code. The
workaround implemented here does a simple (but expensive) string search
in case of an l-value ref and simply replaces the && with just &.