Fixed #64: Implicit lambda captures are captured by copy.
As http://eel.is/c++draft/expr.prim.lambda#capture-10 states, implicit
lambda captures are capture by copy. Insights used the variable type
before which produced an reference for a reference type. This fix strips
the reference in case of an implicit capture and make the resulting
lambda complying  to the standard.