Fixed noexcept specifier evaluation.
A noexcept specifier can either be just the specifier or a noexcept
expression. In the latter case it evaluates to true or false. To get the
expression result it needs another check. Also we can encounter a
CXXNoexceptExpr during parsing which was not evaluated either.

With this change a noexcept specifier is given it it was specified in
the source code. In case of a noexcept expression the expression result
(true/false) is given together with the noexcept specifier to indicate
that the difference.