Fixed crash in case of a PredefinedExpr which is unresolved.
A PredefinedExpr within the definition of a template is unresolved, e.g.
for __func__ the name of the function is unknown in the template itself.
It first appears in an instantiation of a template. Before the fix the
assumption was, that a PredefinedExpr always returns a StringLiteral
with the function name. In the case when the PredefinedExpr was
unresolved that StringLiteral is nullptr. This fix adds a check and
inserts the predefined name in case the name is currently unresolved.