Unified lambda generation.
Previously lambdas had its own generation code. This made it difficult
to show implicit special member functions. Additionally the static
invoker was not generated in all cases. This rewrite moves the lambda
generation into the CXXRecordDecl parsing which known how to create a
class and handle the special member functions. As a additional benefit
it does handle the static invoker correctly, as it knows how to deal
with member functions.

This change also adds support for the mutable keyword attached to a
member variable.