Fixed #91: Add parentheses for BinaryOperator with `+` or `-`.
At least in variadic templates with fold expressions the expanded
version does not carry parentheses. For certain operations like `+` and
`-` this leads to wrong results. This fix adds the required parentheses
in the case we have a BinaryOperator and if it is either a `+` or `-`..

See also N4800 [temp.variadic] p10.1 - p10.2.