Fixed #158: In case of a global braced-init not all chars where deleted.
In case of a global variable with braced-init `findLocationAfterToken`
does not return the location after the semi. Instead it returns a
location at the end of the init-expr. This leaves some characters in the
result. This fix tries to find the right-brace, if a semi is required
which does the job. To move only the location which need it a new
parameter was introduced `RequireSemi` which activates the behaviour for
`FindLocationAfterSemi` and `GetSourceRangeAfterSemi`.