Fixed #205: Make lambdas as in class initializer work.
Lambdas as an in class initializer cause a `LambdaExpr` to occur in a
`CXXConstructorDecl`s initializer list. This causes the lambda to be
placed at a wrong location.

It also triggers a behaviour where the `LambdaExpr` occurs twice. Once
in the in class initializer as written and later in the `CXXConstructorDecl`.
The first one cause a crash due to a `nullptr` access. This patch
filters these lambdas out.