Added support for using libc++ under Linux.
During adding libc++ support for Linux also addressed a Clang
`resource-dir` issue. This patch adds support to figure out the Clang
resource dir and add it to the command line as first arguments. Those a
user should be able to override them. In general, this hopefully solves
a couple of missing include errors.