Fixed #290: Added support for unions and anonymous structs.
To this point unions where shown as a struct, as Clang internally models
them as a CXXRecordDecl.

Also unsupported where anonymous structs. Due to the declaration order
in Clang a name based on line/column prefixed with `__anon_` is made up
to be able to refer to the struct later in a `FieldDecl`.