Fixed #298: Handle additional semi at the end of a function template.
Getting the correct end location of a function template seems to be
challenging. In case of #298 it is followed by an additional semi which
results in an `EmptyDecl`. This fix takes a possible `EmptyDecl` into
account and tries to find the end depending whether a semi is required.
Which is for example in case of CTAD.