Added support for `runTest.py` under Windows.
To achieve this, the binary name in `CMakeLists.txt` is now platform
independent. In addition `cmake` is used to find the name of the `python3`
binary. The Windows binary is now located directly in the build folder
and now longer under `Release` or `Debug`.

To be as independent as possible, differences in test results are shown
via Python `difflib` instead of the former Unix `diff`.