Fixed #311: Added support for attributes in general.
Before this patch C++ Insights did not support attributes. The missing
`alignas` specifier that is the reason for #311 is modeled as an
attribute. Hence, this patch also brings support for all other C++
attributes.

Along these lines, P1143 `constinit` which is also modeled as an attribute, is
supported as well.