Fixed #347: Use `const &` or `&/&&` as constructor parameters for lambdas.
The constructor of a lambda is made up by C++ Insights. Due to that it
is hard to get the constructor right. The Standard requires
direct-initialization. However, the members of a lambda are `private`.
The best way is to keep them `private` and create a constructor. With
this fix the constructor takes the parameters either by `&/&&` or by
`const &`. This reduces additional copies as the previous version
showed. However, cases when variables are moved into a lambda still
generate and additional move constructor call. So far there seems to be
no way around this.