Fixed #381: Show move into std::get for structured bindings.
This "move" was already there in form of an xvalue cast. However, these
are suppressed if the options show all casts is disabled. This fix
enables it always for structured bindings.

Along with the primary fix for #381 this patch also makes the resulting
variables correct. In only a few cases they are really references. Most
of the tome they are copies. This can be check with static_assert and
type-traits to the original code.