Fixed# 386: Make lambdas as single statment of if work.
In the case of an IfStmt, ForStmt or WhileStmt without braces with a
directly invoked lambda as single statement the made-up class was in
the scope of the if but the invoke of the call operator wasn't. That
way the code did not compile and was showing a wrong transformation.

The fix is to wrap _all_ these statements in curly braces, if they are
not already there. This should as a side effect help in showing why
these statements without braces can lead to human errors.