Fixed `noexcept` in destructors.
Destructors are always implicitly inline if not declared i
otherwise. See [except.spec] p8 and [class.dtor] p6. However,
it can be that Clang does not evaluate the exception
specification of a destructor (or any other functions) if
that function is not used.

This patch adds a noexcept comment in this case.