Fixed #520: Avoiding crashing due to unsatisfied requirements.
A `requires` expression can be used as a return value in the body of a
function, for example. In this case, the expression is there for
an instantiation. If now a requirement fails, this instantation is not
satisfied, but there is no printable expression anymore. In this
case, C++ Insights simply creates `requires false` to keep the result.