Use `TranslationUnitDecl` handler instead of the previous matchers.
This change introduces a `TranslationUnit` handler, replacing the previous
matchers. The advantage is that _all_ statements are processed,  not only the
ones where an exact match exists.

The new approach fixes #376, #500, and #588.

There is one disadvantage: for primary templates, some code is not reproduced
as written. This causes one compile issue.